+++ packages/core-data/src/entity-types/icon.ts
@@ -21,6 +21,10 @@ declare module './base-entity-records' {
 			 * The slug of the collection this icon belongs to.
 			 */
 			collection?: string;
+			/**
+			 * Additional search terms for the icon.
+			 */
+			keywords?: string[];
 		}
 	}
 }